Polaczono atrybuty i staty bo karwsz pisze w notatniku i jest uparty
