+++ src/main/java/pl/alvion/rpem/rpessentials/rpplayer/RPPlayer.java
@@ -4,7 +4,6 @@ package pl.alvion.rpem.rpessentials.rpplayer;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
 import pl.alvion.rpem.rpessentials.RPEssentials;
-import pl.alvion.rpem.rpessentials.rpplayer.attribute.Attribute;
 import pl.alvion.rpem.rpessentials.rpplayer.stats.Stats;
 import pl.alvion.rpem.rpessentials.rpplayer.traits.Traits;
 
@@ -27,7 +26,6 @@ public class RPPlayer {
     protected String conigPath;
     private double CurrentFractionHP;
     private ArrayList<Traits> playerTraits = new ArrayList<>();
-    private ArrayList<Attribute> playerAttributes = new ArrayList<>();
     RPPlayer(Player player) {
         this.player = player;
         conigPath = "Player." + this.player.getDisplayName();
@@ -59,9 +57,6 @@ public class RPPlayer {
         return CurrentFractionHP + this.getPlayer().getHealth()*100;
     }
 
-    public ArrayList<Attribute> getPlayerAttributes() {
-        return playerAttributes;
-    }
 
     public void addTrait(Traits trait) {
         Traits.addPlayerTrait(this.player, trait, 0,0);
