--- src/main/java/pl/alvion/rpem/rpessentials/rpplayer/attribute/Attribute.java
@@ -1,10 +1,0 @@
-package pl.alvion.rpem.rpessentials.rpplayer.attribute;
-
-
-
-import java.util.ArrayList;
-
-public abstract class Attribute {
-    public static ArrayList<Attribute> Attributes = new ArrayList<>();
-
-}
