--- src/main/java/pl/alvion/rpem/rpessentials/rpplayer/attribute/AttributeTypes.java
@@ -1,7 +1,0 @@
-package pl.alvion.rpem.rpessentials.rpplayer.attribute;
-
-public enum AttributeTypes {
-    Offence,
-    Crafts,
-    Body;
-}
