+++ src/main/java/pl/alvion/rpem/rpessentials/rpplayer/stats/AttributeTypes.java
@@ -1,0 +1,7 @@
+package pl.alvion.rpem.rpessentials.rpplayer.stats;
+
+public enum AttributeTypes {
+    Offence,
+    Crafts,
+    Body
+}
